Разделил классы с базовы набором модификаторов функций и кастомных
