+++ src/index.php
@@ -2,6 +2,7 @@
 require 'SuQLLog.php';
 require 'TuringMachine.php';
 require 'SuQLEntityHelper.php';
+require 'SQLBaseModifier.php';
 require 'SQLModifier.php';
 require 'SuQLHandler.php';
 require 'SQLBuilder.php';
